updated player component to useRef instead of onChange
